// fe/src/lib/api/generated/model/distributionSection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { DistributionRow } from './distributionRow';

export interface DistributionSection {
  denominator: number;
  key: string;
  label: string;
  rows: DistributionRow[];
}