// fe/src/lib/api/generated/model/programColumn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { OrganizationReference } from './organizationReference';

export interface ProgramColumn {
  campus: OrganizationReference | null;
  code: string | null;
  college: OrganizationReference | null;
  is_legacy: boolean;
  key: string;
  name: string;
  program_id: string | null;
}